apps/api/pyproject.toml
[project]
name = "conflo-api"
version = "0.1.0"
description = "Conflo Construction PM API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.1.0",
    "python-jose>=3.3.0",
    "httpx>=0.26.0",
    "boto3>=1.34.0",
    "stripe>=8.0.0",
    "sentry-sdk[fastapi]>=1.40.0",
    "redis>=5.0.0",
    "resend>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.backends._legacy:_Backend"
